Rework memoization, fixing infinite loop problems
Summary:
See copious comments in the source for details.  The problem with the
previous approach was that it was too easy to lose the continuation that
contained the memo table update when exceptions are flying around, and
losing that continuation always results in an infinite loop.  This
solution is more robust, because we can never lose a continuation, the
worst that could happen (and it shouldn't happen) is that we do a bit of
repeated work.

Test Plan: unit tests, my infinite loop repro case is fixed

Reviewed By: jon.coens@fb.com

Subscribers: ldbrandy, smarlow, akr, bnitka, jcoens

FB internal diff: D1427283

Tasks: 3734942