Introduce possibility of a bounded cache.
There is a new typeclass, CacheableSource, with a single function
cacheSize. This function specifies the size (in terms of the number of
requests) that are cached for each request. 'Nothing' indicates an
unbounded cache, while 'Just n' says that at most n requests are cached.

Bounded caches allow handling of large data volumes, at the expense of
sacrificing the consistency guarantee that the cache usually grants in
Haxl.

The bounded cache is based on the psqueues package.