Attach a call stack to HaxlException
Summary:
When profiling is on (GHC's -prof flag) we can grab a call stack with
GHC.Stack.currentCallStack.  This diff attaches the call stack to
every HaxlException when profiling is on.

(this may have some runtime overhead, but it's only when profiling is
on, and if it turns out to be a problem for accurate profiling then we
can fix it by deferring rendering of the stack until we need to see
it).

Reviewed By: zilberstein

Differential Revision: D2754599

fbshipit-source-id: d71f71c3f897aec7db92a3446cfc653158a1a6f4