Specialised requests
Sometimes, one request is a special case of a more general request.  In
case the general case has already been requested, we can avoid
scheduling the special request, and determine its result from the result
of the more general request.

An example would be determining the number of some xs (the special case)
and retrieving the xs (the general case).