Add new exception category: LogicBug
Summary:
There are some exceptions that are really bugs, such as passing
invalid parameters to functions, or violating invariants.  We want to
fix these bugs, but they are currently represented by exceptions in
LogicError, and are often caught.

Let's make a separate exception category, LogicBug, and start
migrating exceptions over.  This will expose bugs that need fixing, so
it will be a difficult migration.

This is step #1, introduce the new category.  Any new exceptions we
create that indicate client-coe bugs should be a child of LogicBug.

Reviewed By: xich

Differential Revision: D3201590

fbshipit-source-id: 4fd9ed688e8d327b8fe9543dd6998d7efbc46c27