POC: Add native Haxl monad profiling
Summary:
Given that Haskell profiling is really heavyweight and can't be used
in prod, I wanted to experiment with a lightweight profiling scheme in
the Haxl monad itself.

This should let us get per-policy profile data on a sampling basis
from production, but it won't be as detailed as full profiling - no
stack traces - but it does take into account memoization.

The idea is fairly simple:
* There's a notion of a current "label" (a Text).  Policies and
  memoized features are labels.
* Allocations are tracked and assigned to the current label.
* We also track dependencies, when one label depends on another is
  recorded.  This is so that we could account for memoization in
  post-processing of the data.
* We could also assign data fetches to the current label (not done
  yet).

When turned off, this only costs a simple test at each label site
(policy entry and memo point).

If we want to use this, there's a lot to do:
* Allocs include memoized dependencies right now so they aren't
  correct), this is fixable, but it's a bit tricky to do it without
  breaking allocation limits, so I stopped to make a POC diff.
* Measure the overhead, both turned on and off.
* Include data-fetching
* Implement sampling in prod
* Figure out where to put with the data, and post-process it, etc.

Reviewed By: xich

Differential Revision: D3201803

fbshipit-source-id: 623c2efd602fc071ac6d80501474c09c3cda5c18