Add support for disabling the cache
Summary:
Some people want to use Haxl for large batch jobs or long-running
computations.  For these use-cases, caching everything is not
practical, but we still want the batching behaviour that Haxl provides.

The new Flag field, caching, controls whether caching is enabled or
not.  If not, we discard the cache after each round.

Reviewed By: niteria

Differential Revision: D3310266

fbshipit-source-id: 3fb707f77075dd42410bd50fc7465b18b216804e