Enable caching while recording tests for uncachedRequest
Summary:
Try different approaches, but ended up with this. At first I wanted to use SIEnv, but that is not available from Core (Fetches.hs) and we don't want that dependency.
Then I tried creating a haxlsh command for it, but I didn't like it as it will be not useful. So I pass it via the haxlConfig and set the env just before running the test.

The only thing that I am not sure about is if we need to deprecate uncachedResponse. The new signature is a more restricted type that implements the Show type class. As this
was already called with responses everywhere, it seems to work out of the box.

I am a bit confused though for the BlackBird issue. Memcache will fail Set, but shoudn't fail Get as that is using dataFetch which is cached already. It might have been another issue.

Reviewed By: simonmar

Differential Revision: D6195542

fbshipit-source-id: 483d7ec4f32a27b8c9cbfb5d9b129abe40a93c54