log to sigma_datasources scuba table
Summary:
It is useful to have a separate scuba table to record datasource stats where the datasource name is a row instead of a column. This allows us to do more complex aggregations over time/allocations/etc of ALL datasources not per each individual one.

The way it's set up now I rate limit on the datasource name. I'm not sure if this is the right move, maybe I should rate limit on both the datasource and the context? I also log each fetch as a separate row. This might end up being expensive, but it allows us to get more detailed information about batch sizes (ie p95, avg, sum). If we combined all fetches for each datasource then we'd get less detailed information, but it may be more practical.

Reviewed By: sameenjalal, simonmar

Differential Revision: D6492647

fbshipit-source-id: 28399ece277f93616786bdfb18264a0250aa238b