- Use SourceCode Location to Calculate Memoization Labels
Summary:
Use the source location of the `memo` statement to calculate the memoization label as opposed to the memoization key. This is to make sure that multiple `memo` calls in the same declaration will get different labels.

This enables to remove `memoUniqueText` and the related pragma as `memoUnique` will now have a Text label by default.

Reviewed By: simonmar

Differential Revision: D9779196

fbshipit-source-id: 89d2ce6d5a6ebc4a6b8246a5ef561ecf8e18c522