Collect allocation stats for background fetch
Summary:
Right now, we get no alloication infroamtion for background fetch at all.
This is because the allocation usually happen in a thread and we only get
back the control in a callback. As a improvement, we can just use the total
allocation in the child thread as the allocation.

While individual allocation won't be correct if `putResult` is called more than
once. The total sum and amortized allocation will still be helpful stats.

Reviewed By: simonmar

Differential Revision: D14734588

fbshipit-source-id: 6beed3741a496d308bdcec75eb0f0e0bbfbbae38