Reify getIVar into Cont type
Summary:
Reify the getIVar continuation into the Cont type.

There are two cases where the continuation of a Blocked is getIVar:

  * A data fetch
  * When getIVar itself is blocked

Being able to spot the case where the continuation is a simple getIVar call
allows us to avoid creating an extra IVar sync point in the definition of
`<*>`. We can just use the existing IVar as the sync point.

Reviewed By: josefs

Differential Revision: D15587815

fbshipit-source-id: 11f354495bfc0789d292963f4786986b335fc4f1