Special case g :<$> Return i in blockedBlocked
Summary:
Special-case the handling of fmap-over-return in the blocked computation on the left of `<*>`. This mainly helps the two common cases of `sequence` and `traverse`.

Consider `sequence [x,y,z]`. It generates an applicative tree like:

```
            <*>
           /   \
  (:) <$> x     <*>
               /   \
      (:) <$> y     <*>
                   /   \
          (:) <$> z     pure []
```

Without this special case, `blockedBlocked` will fall through to its last case, which creates an extra IVar for the join. With this case, we can look through the `:<$>` to see if `x` is a `Return`. If so, we can avoid the IVar creation.

Reviewed By: simonmar

Differential Revision: D15689684

fbshipit-source-id: 5e09e70d99c5791710963790e6fd3cbd43c5fab7