rethrow async exceptions from unsafeToHaxlException
Summary: Haxl is not async exception safe, in that the environment could be left in a bad state if the exception was thrown at just the wrong time. Rahter than fixing this right now which would be a large project, we just rethrow all async exceptions out of the computation. Since there is a chance the exception could have been thrown in the scheduler anyway, this is unlikely to change behaviour in a meaningful way.

Reviewed By: simonmar

Differential Revision: D19410092

fbshipit-source-id: e6f4b082cb854c517d174fdc8ddba8fa10f169b3