Include callstack in the exception from data cache hit
Summary:
Currently, we only provide callstack if the exception is thrown in Haxl in
user code (by calling `Haxl.throw`). There is not callstack in the
`HaxlException` if the exception origins from data fetches or cached result.

In this diff, we added such callstack information when we get the exception
from cache hit. Currently, it still doesn't work if we get an exception from
the data source fetch itself. Ideally this is what we want. But we can still
get meaningful debugging information by running a function, keeping the cache
and rerunning it.

Reviewed By: josefs

Differential Revision: D19650710

fbshipit-source-id: fcb8a16b94af40bfa1a4064052ffac4d57d42532