Use record syntax for IVar
Summary:
Simply add record definition and getter for newtype IVar.
The only behavior change is that we added bang patterns for a few cases where
the IORef is already used. This was done for most cases before but missed in
a few places.
This almost no-op change will make it easier to change the definition of IVar
in profiled version in stacked change.

Reviewed By: josefs

Differential Revision: D19855621

fbshipit-source-id: eddbcba48c83f5ba9eb8b1ea312760568c390fd5