Include stacktrace information in the exception from data fetches
Summary:
In profiled rts, we have stacktrace information when user throws an haxl
exception in Haxl monad. But we don't have such information when a data fetch
throws an exception as data fetch happens in IO and in the scheduler.
In this change, We capture the CCS when scheduling a data fetch, and populate
it back when an exception is thrown. As a result, we will be able to see
relavant callstack in a data fetch failure.

Reviewed By: josefs

Differential Revision: D19855626

fbshipit-source-id: d266016555fcc5299269591cc2b0c91aea348638