profile more accurately (#119)
Summary:
Pull Request resolved: https://github.com/facebook/Haxl/pull/119

Currently profiling with labels does not track full stacks.
This is problematic in the case where all top level methods end up calling a common method with a label, as we have no way of attributing to the top level methods.
Eg if A/B/C all call X when running work then we cannot tell which work is expensive, as we have no connection from the work through X

Reviewed By: simonmar

Differential Revision: D20384255

fbshipit-source-id: f9aa0462904c17dee32d37a659b491e8d252d6db