remove speculative/imperitive (#125)
Summary:
Pull Request resolved: https://github.com/facebook/Haxl/pull/125

These are confusing and make the behaviour of pAnd/pOr confusing.
In general you might expect these to run both sides of the branch at the same time.
In fact it will run each branch concurrently, but slower (by design) in order to limit the amount of computation that gets done in case it is terminated early.

This results in performance hits if you convert an && to a pAnd when both branches are highly likely to be true, which is the *exact* case to use pAnd.

Reviewed By: simonmar

Differential Revision: D22410467

fbshipit-source-id: f1536afd80acf6b6c527b7618fdcf510b87735c2