Add function to sanitize the env in case of exceptions
Summary: It is possible for a Haxl computation to end with a failed result rendering the Env unsafe for use in future computations. This diff adds a function `sanitizeEnv` which removes all empty IVars from the cache, and clears the different queues the scheduler maintains in the Env - so that it can be reused.

Reviewed By: simonmar

Differential Revision: D22397982

fbshipit-source-id: c593a63747d61a56df9a910da22ae1789abc80c5