supoprt classifying exceptions as ignored for stats
Summary:
Some datasources will throw exceptions but this does not indicate a problem with the datasource itself. This can make the statistics difficult to use to track actual problems versus problems with the way the datasource is used.
Here we allow the datasource to classify some failures to be ignored by the stats collection. They are not simply ignored however - but stored in a new field `fetchIgnoredFailures`

Reviewed By: josefs

Differential Revision: D23475953

fbshipit-source-id: 9c6d5e81d00e2665ace84f089bd272c9d566eb3c