Add time to profile
Summary:
To track down bugs where a single label is taking a long time, but is not allocating - it is useful to store wall clock time on labels.
This is obviously only an estimate, due to the lazy nature of Haskell - but it will at least give a clue to where to start looking.

Reviewed By: anubhav94N

Differential Revision: D24083517

fbshipit-source-id: e6058a227345b5a51a7ef05f5369803f91bd61c4