Add time to profile (#131)
Summary:
Pull Request resolved: https://github.com/facebook/Haxl/pull/131

To track down bugs where a single label is taking a long time, but is not allocating - it is useful to store wall clock time on labels.
This is obviously only an estimate, due to the lazy nature of Haskell - but it will at least give a clue to where to start looking.

Reviewed By: anubhav94N

Differential Revision: D24083517

fbshipit-source-id: e778c1c9125104c53935bb7ecc2ab0d6134e386c