Make report flags a bitmask instead of sequential numbers
Summary:
Make report flags a bitmask instead of sequential numbers. The reporting
behavior may not be strictly ordered, some reporting behavior can be enabled
independently. By not using magic numbers, it will also be easier to introduce
new reporting behavior.

We plan to add a new `ReportExceptionStack` flag include label stack in
`HaxlException` in the future, which can be enabled independently with
`ReportFetchStack`.

Reviewed By: simonmar

Differential Revision: D32933603

fbshipit-source-id: 4028aa91b68caa738684a77392c916970b4605c9