docs: add GH button in support of Ukraine (#147)
Summary:
Our mission at [Meta Open Source](https://opensource.fb.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

Pull Request resolved: https://github.com/facebook/Haxl/pull/147

Reviewed By: simonmar

Differential Revision: D34649194

Pulled By: dmitryvinn-fb

fbshipit-source-id: 6bb709a7ed5587667f13aadd7facba04ca9d3546